Update inthook.py drop python 2 support
"This is our inthook implementation: we do it to support both PY2 and PY3 in GDB. 
Why? 
Because int(gdb.Value(...)) on GDB with PY2 didn't work.
So we hooked int globally via globals()['int'] = xint and globals()['long'] = xint (in py2) 
and handled "proper conversions to int" by ourselves.

This code has been changed through time. 
e.g. add support for Python 3 enum classes or to fix some bugs that were found in the meantime."